+++ project/templates/login.html
@@ -24,7 +24,7 @@ body {
 
 
 {% block content %}
-<form class="form-signin" action="{{ url_for('login') }}" method="POST">
+<form class="form-signin" action="{{ url_for('auth.login') }}" method="POST">
     <img class="mb-4" src="/static/images/dummy.png" alt="" width="72" height="72">
     <h1 class="h3 mb-3 font-weight-normal">Please login</h1>
