+++ project/templates/signup.html
@@ -25,7 +25,7 @@ body {
 
 
 {% block content %}
-<form class="form-signin" action="{{ url_for('signup') }}" method="POST">
+<form class="form-signin" action="{{ url_for('auth.signup') }}" method="POST">
     <img class="mb-4" src="/static/images/dummy.png" alt="" width="72" height="72">
     <h1 class="h3 mb-3 font-weight-normal">Please signup</h1>
